bindings: (partial fix #2451) Handle std::_Ios_Openmode typedef
std::_Ios_Openmode is a typedef of std::ios_base::openmode bitmask type.
This mapping needs to be provided to the type handler.  The previous
mapping was for std::ios::openmode (ios is a subclass of ios_base),
but the use of subclass here no longer works.  Thanks to Gustavo
Carneiro for a hint of where to address this issue in the bindings.